App.js updated
Take in mind that we have a big change here: the old index.ios.js and index.android.js files were removed to give it place to index.js and App.js.
These two files have the same effect of those older two. However, here we must  understand that at the end of App.js the "expor default App;" line
must be added so that other parts of the Application can use it.